* Add Let's Encrypt root self-signed certificate to chain to be installed.
VCSA certificate-manage returns `Unable to validate the chain of trust for the provided SSL certificate and Root.` when deploying Let's Encrypt certificate. This is caused by Self Signed root missing on fullchain and ca-certificates of VCSA do not have it on OpenSSL chain. So the root self-signed certificate must be on fullchain too.